@@ -50,6 +50,7 @@
     "colors": "^1.1.2",
     "cookie-parser": "^1.4.1",
     "css-loader": "^0.23.1",
+    "enzyme": "^2.3.0",
     "express": "^4.13.3",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
@@ -68,11 +69,11 @@
     "node-sass": "^3.4.2",
     "postcss-loader": "^0.8.0",
     "proxy-middleware": "^0.15.0",
-    "react": "^0.14.8",
-    "react-addons-css-transition-group": "^0.14.7",
-    "react-addons-test-utils": "^0.14.7",
+    "react": "^15.0.2",
+    "react-addons-css-transition-group": "^15.0.2",
+    "react-addons-test-utils": "^15.0.2",
     "react-breadcrumbs": "^1.3.10",
-    "react-dom": "^0.14.7",
+    "react-dom": "^15.0.2",
     "react-hot-loader": "^1.3.0",
     "react-input-color": "^1.7.2",
     "react-redux": "^4.4.0",
