@@ -1,8 +1,13 @@
 import React from 'react'
 
-const AppDashboardContents = ({ children }) =>
-  <div className='dashboard-apps'>{ children }</div>
+const AppDashboardContents = ({ children, className = `` }) =>
+  <div className={ `dashboard-apps ${className}` }>
+    { children }
+  </div>
 
-AppDashboardContents.propTypes = { children: React.PropTypes.node.isRequired }
+AppDashboardContents.propTypes = {
+  children: React.PropTypes.node,
+  className: React.PropTypes.string
+}
 
 export default AppDashboardContents
