@@ -11,8 +11,8 @@ import Fieldset from '../elements/Fieldset.component'
 import Row from '../elements/Row.component'
 import FeedbackInput from '../elements/FeedbackInput.component'
 
-const AppForm = ({ submit, name }) =>
-  <AppSection>
+const AppForm = ({ className = ``, name, submit }) =>
+  <AppSection className={ className }>
     <ContextHeader>
       <ContextTitle>Create a New App</ContextTitle>
       <ContextMenu />
@@ -34,7 +34,7 @@ const AppForm = ({ submit, name }) =>
                              value={ name }
                              className='easy'
                              autoFocus='true' />
-              <div>
+              <div className='app-form-name-preview'>
                 { name.value && <span>{ kebabCase(name.value) }</span> }
               </div>
             </Column>
@@ -49,8 +49,9 @@ const AppForm = ({ submit, name }) =>
   </AppSection>
 
 AppForm.propTypes = {
-  submit: React.PropTypes.func.isRequired,
-  name: React.PropTypes.object.isRequired
+  className: React.PropTypes.string,
+  name: React.PropTypes.object.isRequired,
+  submit: React.PropTypes.func.isRequired
 }
 
 export default AppForm
