@@ -1,8 +1,13 @@
 import React from 'react'
 
-const AppSection = ({ children }) =>
-  <section className='app'>{ children }</section>
+const AppSection = ({ children, className = `` }) =>
+  <section className={ `app ${className}` }>
+    { children }
+  </section>
 
-AppSection.propTypes = { children: React.PropTypes.node.isRequired }
+AppSection.propTypes = {
+  children: React.PropTypes.node,
+  className: React.PropTypes.string
+}
 
 export default AppSection
