@@ -6,15 +6,14 @@ import ContextMenu from '../elements/ContextMenu.component'
 import ContextFooter from '../elements/ContextFooter.component'
 import DestroyButton from '../elements/DestroyButton.component'
 
-const AppView = ({ app, deleteApp, opacity }) =>
-  <div className='app-view' style={ { opacity } }>
+const AppView = ({ app, className = ``, deleteApp, opacity = 0.5 }) =>
+  <div className={ `app-view ${className}` } style={ { opacity } }>
     <ContextHeader>
       <ContextTitle>
         { app ? app.name : '...' }
       </ContextTitle>
 
-      <ContextMenu>
-      </ContextMenu>
+      <ContextMenu />
     </ContextHeader>
 
     <div>
