@@ -1,4 +1,5 @@
 import React from 'react'
+import { fromJS } from 'immutable'
 import App from './App.container'
 import AppDashboard from './AppDashboard.component'
 import AppDashboardContents from './AppDashboardContents.component'
@@ -13,20 +14,30 @@ export default class Apps extends React.Component {
   static propTypes = {
     apps: React.PropTypes.object.isRequired,
     addApp: React.PropTypes.func.isRequired,
+    className: React.PropTypes.string,
     invalidateCache: React.PropTypes.func.isRequired
   }
 
-  componentWillMount() { this.props.invalidateCache() }
+  static defaultProps = {
+    apps: fromJS([]),
+    addApp: event => event.preventDefault(),
+    className: ``,
+    invalidateCache: () => {}
+  }
+
+  componentWillMount() {
+    this.props.invalidateCache && this.props.invalidateCache()
+  }
 
   render() {
-    const { apps, addApp } = this.props
+    const { apps, addApp, className } = this.props
     const onlyOneApp = apps.count() === 1
     const noApps = apps.count() === 0
     const hasDashboard = apps.getIn([0, 'name']) === 'supergiant'
     const isDefaultInstall = (onlyOneApp && hasDashboard) || noApps
 
     return(
-      <AppDashboard>
+      <AppDashboard className={ className }>
         <ContextHeader>
           <ContextTitle>My Apps</ContextTitle>
 
