@@ -1,7 +1,6 @@
 import { takeEvery } from 'redux-saga'
 import { call, fork, put } from 'redux-saga/effects'
 import { push } from 'react-router-redux'
-import { getCache } from './apps.selectors'
 import * as Resources from '../resources'
 import * as Actions from './apps.actions'
 import {
@@ -11,9 +10,9 @@ import {
 
 const Client = new Resources.Apps
 
-function* fetchApps(getState) {
+export function* fetchApps(getState) {
   try {
-    let apps = yield call(Client.fetch)
+    let apps = yield call([ Client, Client.fetch ])
 
     for (let app of apps) { yield put(Actions.insert(app.key(), app)) }
   } catch (error) {
@@ -21,7 +20,7 @@ function* fetchApps(getState) {
   }
 }
 
-function* invalidateCache(getState, { id }) {
+export function* invalidateCache(getState, { id }) {
   if (id) {
     yield fork(getApp, getState, { id })
   } else {
@@ -29,9 +28,9 @@ function* invalidateCache(getState, { id }) {
   }
 }
 
-function* getApp(getState, { id }) {
+export function* getApp(getState, { id }) {
   try {
-    let app = yield call(Client.get, id)
+    let app = yield call([ Client, Client.get ], id)
 
     yield put(Actions.insert(app.key(), app))
   } catch (error) {
@@ -39,9 +38,9 @@ function* getApp(getState, { id }) {
   }
 }
 
-function* createApp(getState, { params }) {
+export function* createApp(getState, { params }) {
   try {
-    let app = yield Client.create(params)
+    let app = yield call([ Client, Client.create(params) ])
 
     yield fork(infoMessages, `Created '/${app.name}'`)
     yield put(push(`/`))
@@ -50,9 +49,9 @@ function* createApp(getState, { params }) {
   }
 }
 
-function* deleteApp(getState, { id }) {
+export function* deleteApp(getState, { id }) {
   try {
-    yield call(Client.delete, id)
+    yield call([ Client, Client.delete ], id)
     yield put(Actions.remove(id))
     yield fork(infoMessages, `Deleted app ${id}`)
     yield put(push(`/`))
@@ -61,23 +60,23 @@ function* deleteApp(getState, { id }) {
   }
 }
 
-function* fetch(getState) {
+export function* fetch(getState) {
   yield takeEvery(Actions.Fetch, fetchApps, getState)
 }
 
-function* get(getState) {
+export function* get(getState) {
   yield takeEvery(Actions.Get, getApp, getState)
 }
 
-function* create(getState) {
+export function* create(getState) {
   yield takeEvery(Actions.Create, createApp, getState)
 }
 
-function* destroy(getState) {
+export function* destroy(getState) {
   yield takeEvery(Actions.Destroy, deleteApp, getState)
 }
 
-function* invalidate(getState) {
+export function* invalidate(getState) {
   yield takeEvery(Actions.Invalidate, invalidateCache, getState)
 }
 
