@@ -14,7 +14,7 @@ export default class Releases {
   constructor({ client, scope }) {
     const api = composeApi({ client, scope })
 
-    for (let property of Reflect.ownKeys(api)) {
+    for (let property of api) {
       let method = Reflect.get(api, property)
 
       Reflect.set(this, property, method)
