@@ -87,32 +87,47 @@ export default class ResourceClient {
     this.resource = resource
   }
 
-  get = key => {
+  *[Symbol.iterator]() {
+    let publicInterfaces = [
+      `client`,
+      `create`,
+      `delete`,
+      `fetch`,
+      `get`,
+      `resource`,
+      `route`,
+      `update`
+    ]
+
+    for (let publicInterface of publicInterfaces) { yield publicInterface }
+  }
+
+  get(key) {
     let handler = this.client ? this.client.getResource : getResource
     let get = handler(this.route)
 
     return get(key)
       .then(throwErrorUnless(200))
       .then(marshalResource(this.resource))
-  };
+  }
 
-  fetch = () => {
+  fetch() {
     let handler = this.client ? this.client.getResources : getResources
     let fetch = handler(this.route)
 
     return fetch()
       .then(throwErrorUnless(200))
       .then(marshalCollection(this.resource))
-  };
+  }
 
-  delete = key => {
+  delete(key) {
     let handler = this.client ? this.client.deleteResource : deleteResource
     let destroy = handler(this.route)
 
     return destroy(key).then(throwErrorUnless(202))
-  };
+  }
 
-  create = params => {
+  create(params) {
     let resource = new this.resource(params)
     let handler = this.client ? this.client.createResource : createResource
     let create = handler(this.route)
@@ -120,14 +135,14 @@ export default class ResourceClient {
     return create(resource.toPayload())
       .then(throwErrorUnless(201))
       .then(marshalResource(this.resource))
-  };
+  }
 
-  update = resource => {
+  update(resource) {
     let handler = this.client ? this.client.updateResource : updateResource
     let update = handler(this.route)
 
     return update(resource.key(), resource.toPayload())
       .then(throwErrorUnless(202))
       .then(marshalResource(this.resource))
-  };
+  }
 }
