@@ -1,7 +1,6 @@
 import { takeEvery } from 'redux-saga'
 import { call, fork, put } from 'redux-saga/effects'
 import { push } from 'react-router-redux'
-import { getCache } from './tasks.selectors'
 import * as Resources from '../resources'
 import * as Actions from './tasks.actions'
 
@@ -14,7 +13,7 @@ const Client = new Resources.Tasks
 
 function* fetchTasks(getState) {
   try {
-    let tasks = yield call(Client.fetch)
+    let tasks = yield call([ Client, Client.fetch ])
 
     for (let task of tasks) {
       yield put(Actions.insert(task.key(), task))
@@ -34,7 +33,7 @@ function* invalidateCache(getState, { id }) {
 
 function* getTask(getState, { id }) {
   try {
-    let task = yield call(Client.get, id)
+    let task = yield call([ Client, Client.get ], id)
 
     yield put(Actions.insert(task.key(), task))
   } catch (error) {
@@ -44,7 +43,7 @@ function* getTask(getState, { id }) {
 
 function* deleteTask(getState, { id }) {
   try {
-    yield call(Client.delete, id)
+    yield call([ Client, Client.delete ], id)
     yield put(Actions.remove(id))
     yield put(Actions.invalidate(id))
     yield fork(infoMessages, `Deleted task ${id}`)
